Move example to examples/basic
