+++ examples/basic/public/index.html
