+++ examples/basic/src/server/schema.js
