+++ package.json
@@ -18,8 +18,8 @@
     "clean-types": "rimraf types",
     "build-types": "npm run clean-types && tsc -p tsconfig.dts.json",
     "build": "builder concurrent --buffer build-lib build-es build-types",
-    "demo:start:graphql-server": "node example/src/server/index.js",
-    "demo:start:app": "webpack-dev-server --hot --inline --config \"example/webpack.config.js\"",
+    "demo:start:graphql-server": "node examples/basic/src/server/index.js",
+    "demo:start:app": "webpack-dev-server --hot --inline --config \"examples/basic/webpack.config.js\"",
     "demo:start": "builder concurrent demo:start:app demo:start:graphql-server",
     "type-check": "tsc",
     "lint": "tslint 'src/**/*.{ts, tsx}'",
